backend/requirements.txt
aiosmtplib==3.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asyncpg==0.31.0
azure-appconfiguration==1.8.0
azure-core==1.38.3
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.3
azure-monitor-opentelemetry==1.8.7
azure-monitor-opentelemetry-exporter==1.0.0b49
azure-servicebus==7.14.3
black==24.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
msal==1.35.1
msal-extensions==1.3.1
msrest==0.7.1
mypy==1.9.0
mypy_extensions==1.1.0
oauthlib==3.3.1
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-psycopg2==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-urllib==0.61b0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
ruff==0.3.2
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==1.17.3
zipp==3.23.0
